backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
python-multipart>=0.0.9
# Removed jq as it fails to build on Windows and is unused
# Removed pandas/numpy/boto3/motor as they are unused in this simplified implementation
pytest>=8.0.0
requests>=2.31.0
